deepen: inline queueWelcomeEmail into onboarding action
`lib/email/welcome-service.ts` was a 91-LOC module wrapping one
`setImmediate` fire-and-forget send with one production caller
(`app/(auth)/onboarding/actions.ts:451`). The interface — three
positional args, `Promise<void>` return — earned no leverage: nothing
composed with it, nothing tested it in isolation (no dedicated test
file), and the only caller's `.catch()` handler on the returned
promise was already dead code because the send happens inside
`setImmediate`.

Move `queueWelcomeEmail` into `actions.ts` as a module-private
function alongside its sole caller. The [Onboarding] completion flow
now reads top-to-bottom: DB writes, Clerk metadata sync, welcome
email, first-filing delivery, Stripe reconcile — all in one file.

Locality: the fire-and-forget welcome email lives inside the "user
just completed onboarding" flow rather than one directory over. The
`SecureEmailLogger` and `FOUNDER_REPLY_TO` imports still cross real
seams (secure logging and the founder-reply-to constant are shared
with other email paths).

Test surface: the reply-to regression suite is a source-grep locking
in `replyTo: FOUNDER_REPLY_TO` on the welcome path. Repoint the grep
at the new location (`app/(auth)/onboarding/actions.ts`); all 4
regression tests still pass. Also fix a stale docstring reference in
`lib/email/config.ts:9`.
